+++ R/app_server.R
@@ -5,5 +5,18 @@
 #' @import shiny
 #' @noRd
 app_server <- function(input, output, session) {
-  # Your application server logic
+    # --- Data
+    r_data <- reactive({mtcars})
+
+    # --- Select ---
+    mod_eda_select_server(
+        id = NULL,
+        r_data = r_data
+    )
+
+    # --- Freq table ---
+    mod_eda_freq_table_server(
+        id = NULL,
+        r_data = r_data
+    )
 }