+++ NEWS.md
@@ -1,3 +1,12 @@
+# shimo.eda 0.0.0.9002 (2022-05-14)
+
+Bugfix mod_eda_select.R
+
+- Fixed title bug for data table in module `mod_eda_select.R`
+- Added server utility `tabitem_vertical_space()`
+
+----------
+
 # shimo.eda 0.0.0.9001 (2022-05-14)
 
 mod_eda_select